requirements.txt
asgiref==3.3.1
Django==3.0.4
djangorestframework==3.10.3
psycopg2-binary==2.8.5
pytz==2020.5
sqlparse==0.4.1
